[dev8] Address sequence parser review feedback
