+++ packages/dev8/README.md
@@ -1,17 +1,25 @@
 # Developer Connection
 
-This script is bundled into Studio development builds so the app can communicate with 
+This script is bundled into Studio development builds so the app can communicate with dev8.
 
 ## Development version build
-Run a webpack dev server that auto rebuild your dev8.js on save.
+
+Run a webpack dev server that automatically rebuilds `dev8.js` on save.
 
 ```bash
 npm install
 npm run serve
 ```
 
+The development bundle is hosted at:
+
+```text
+https://localhost:9000/dev8.js
+```
+
 ## Production version build
-The script can be built from this repository by running
+
+The script can be built from this repository by running:
 
 ```bash
 npm install
@@ -24,19 +32,26 @@ npm run build
 npm test
 ```
 
-## Using a locally build version on a project
+## Using a local dev8 build in Studio
 
-Next, find your local ip by running something like
+1. Start the local dev8 development server:
 
-```bash
-ifconfig en0
-```
+   ```bash
+   npm run serve
+   ```
 
-Then go to a cloud editor project and add this line to `head.html`:
+2. Confirm that `https://localhost:9000/dev8.js` is being hosted.
 
-```html
-<script src="https://10.8.8.[your-ip]:8888/dev8.js">
-```
+3. Open a project in Studio.
 
-Next, navigate to `https://10.8.8.[your-ip]:8888/dev8.js` in the browser on your phone to accept
-https certificates.
+4. In the Studio project, open `config/webpack.config.js` and replace the `createDev8Plugin` line with:
+
+   ```js
+   createDev8Plugin({src: 'https://localhost:9000/dev8.js'}),
+   ```
+
+5. In Studio, open the **System** log tab and click **Restart server**.
+
+6. Refresh the simulator. It should now use the locally hosted dev8 build.
+
+   To confirm that the local build is being used, you can temporarily add an `alert()` to the local dev8 source, refresh the simulator, and verify that the alert appears.
