+++ packages/dev8/package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "build": "./node_modules/.bin/webpack",
     "clean": "rm -rf dist",
-    "test": "node --test test/*-test.js",
+    "test": "mocha --require tsx/cjs test/*-test.ts --reporter=dot",
     "serve": "./node_modules/.bin/webpack-dev-server --devtool eval --mode development --host=0.0.0.0",
     "postinstall": "npm run capnp:prepare; patch-package",
     "capnp:prepare": "./build-capnp-ts.sh"
@@ -16,17 +16,22 @@
     "@babel/core": "^7.4.5",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-typescript": "^7.18.6",
+    "@types/chai": "^5.2.3",
+    "@types/mocha": "^5.2.7",
     "babel-loader": "^8.0.6",
     "base64-js": "^1.5.1",
     "capnp-ts": "^0.2.4",
+    "chai": "^4.5.0",
     "css-loader": "^7.1.1",
     "file-loader": "^6.2.0",
     "html-loader": "^0.5.5",
+    "mocha": "^9.2.2",
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
     "style-loader": "^4.0.0",
     "terser-webpack-plugin": "^5.3.10",
     "ts-essentials": "^9.4.1",
+    "tsx": "^4.23.12",
     "typescript": "^5.8.2",
     "webpack": "^5.106.2",
     "webpack-cli": "^5.1.4",
